restructure testselection
