+++ localstack/testing/testselection/opt_in.py
@@ -1,4 +1,9 @@
+"""
+During initial rollout of the test selection in PRs this is additionally limited to the opt-in list below.
+
+A test selection will only be effective if all files that would be selected under the testselection also have at least one match in the list below.
+"""
+
 OPT_IN = [
     "localstack/services/lambda_",
-
 ]
